// src/app/models/QuickLookup.ts
export class QuickLookup {
  first?: string;
  last?: string;
  destination?: string;
  hasLink?: boolean;
  hasMoreLinks?: boolean;
  id?: number;
  carLength?: number;
  routeFare?: string;
  origTimeMin?: string;
  destTimeMin?: string;
  destinationStationName?: string;
  firstStationName?: string;
  trainHeadStation?: string;
  connectingStationName?: string;
  finalConnectingStationName?: string;
  finalDestinationStationName?: string;
  estDepartureFlag?: number;
  estDeparture?: any;
  constructor() {}
}